hopefully this commit works
Changes:
Pages created for "Home", "About Us", "Get Involved", and "Learn More". All listed pages including "Contact Us" are reachable via the navbar. Style sheet configured to mimic the wireframe as close as possible.

For Me To Do:
The Donate page still needs to be created and connected to the navbar

Concerns/ Queries:
Unsure of how the styling was desired regarding the margins on the page. they can easily be removed if it's not what we wanted. Another thing was that I decided to make the donation and language buttons highlight black if hovered over. This may need to be changed as I assume Evans reason for making the donation button black was to draw attention to one of the most important aspects of the site. so just let me know