Events Page + Volunteer Section
Events page and volunteer page completed.

Bugs:
donate button on navbar doesn't function while on the events page

temporary data:
all of the data in the carousel is completely temporary, all the links go to 1 page. Not entirely sure what to do with this section as the client will want to update this carousel but I know we were wanting to stray away from a database so I'm not sure where to go from here.